Action commands API overhaul (#59)
* Rename bevy_motiongfx to motiongfx

* Changed action generics' location

* Remove color_palette

* Use smallvec & changed ActionMeta to ActionSpan

* Add new api for creating animation sequences

* Removed unused interp files

* Remove unused ActionSpan methods

* Fix docs

* Add stringify_field for consistent output

* Fix action builder test

+++ crates/motiongfx_engine/src/slide.rs
@@ -100,7 +100,7 @@ pub fn create_slide(mut sequences: Vec<Sequence>) -> SlideBundle {
 
     let mut start_time = 0.0;
     for (s, sequence) in sequences.iter_mut().enumerate() {
-        sequence.set_slide_index(s);
+        sequence.set_slide_index(s as u32);
         start_times.push(start_time);
 
         start_time += sequence.duration();
